Implemented ReplicationQueuesHBaseImpl that tracks WAL offsets and replication queues in an HBase table. Only wrote the basic tracking methods, have not implemented claimQueue() or HFileRef methods yet. Wrote a basic unit test for ReplicationQueueHBaseImpl that tests the implemented functions on a single Region Server
Changes requested by Matteo. Checked the existence of the Replication Table before trying to create it and made ReplicationQueuesHBaseImpl more thread-safe by switching over the ordering of certain checks and changing the map to a concurrent-safe one.

Move initialization of the Replication Table to the RS side and rely purely on HBase table for queueId -> row key mapping.

Got rid of importing packages

Modified queueIdToRowKey to correctly work with non-unique Queue-Id's and to also filter out results